deploy 5fcacf69caa3c009f535887ff14a8f8243f67750 (Final delay update (so far, I mean unless there is feedback :) ) (#82)
* Rename Learning_delays_v01.ipynb to Learning_delays.ipynb

Rename file

* Delete Learning_delays.ipynb

* Add files via upload

I have collected most of the delay related functions and variables to a DelayLayer class. Applied surrogate update to delays as suggested by (Tomas Fiers, Markus Ghosh and Alessandro Galloni). Method B was applied. The delay is applied only to the first layer. Appling the delay with Method B subtracts a duration=delay_max from the input duration. So every subsequent push through a layer decreases the duration(or effective duration). So, for applying multiple delay layers, the duration of the inputs needs to be appended with delay_max*number_delay_layer.

* Update Learning_delays.ipynb

* Update Learning_delays.ipynb

* Update Learning_delays.ipynb

* Update Learning_delays.ipynb

* Update Learning_delays.ipynb

* Update Learning_delays.ipynb

* Update Learning_delays.ipynb

* Update Learning_delays.ipynb

* 'Major_delay_learning_update'

* Final delay update) 5fcacf69caa3c009f535887ff14a8f8243f67750